crates/sel4-hal-adapters: Split into modules
